# mapaviajero-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel App API",
        "description": "Documentación de la API Travel App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/paises": {
            "get": {
                "tags": [
                    "Países"
                ],
                "summary": "Listar todos los países",
                "description": "Obtiene la lista de todos los paises disponibles. Accesible tanto para usuarios registrados como no registrados",
                "operationId": "a714e128ee0fec14562fd0e54ca6e31f",
                "responses": {
                    "200": {
                        "description": "Lista de países",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Pais"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Países"
                ],
                "summary": "Crear un nuevo país",
                "description": "Solo accesible para administradores y superadministradores. Permite crear y añadir un nuevo país a la Base de Datos.",
                "operationId": "storePais",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaisDB"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "País creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Pais"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado: solo administradores"
                    },
                    "422": {
                        "description": "Error de validación en los datos enviados"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/paises/{id}": {
            "get": {
                "tags": [
                    "Países"
                ],
                "summary": "Mostrar un país específico (por id)",
                "description": "Obtiene los detalles de un país por su ID. Accesible tanto para usuarios registrados como no registrados",
                "operationId": "showPais",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del país",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "País encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Pais"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "País no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "País no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Países"
                ],
                "summary": "Actualizar un país",
                "description": "Permite modificar los datos de un país. Solo accesible para administradores y superadministradores.",
                "operationId": "updatePais",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del país a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaisDB"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "País actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Pais"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo el administrador puede editar un país"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "País no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "País no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El campo nombre ya existe."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Países"
                ],
                "summary": "Eliminar un país",
                "description": "Permite eliminar un país. Solo accesible para administradores y superadministradores.",
                "operationId": "destroyPais",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del país a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "País borrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "País borrado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo el administrador puede borrar un país"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "País no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "País no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pais": {
                "title": "Pais (Recurso API)",
                "description": "Respuesta de la API para un país, con relaciones y campos amigables",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Argentina"
                    },
                    "bandera_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/bandera.png",
                        "nullable": true
                    },
                    "capital": {
                        "type": "string",
                        "example": "Buenos Aires"
                    },
                    "continente": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 3
                            },
                            "nombre": {
                                "type": "string",
                                "example": "Europa"
                            }
                        },
                        "type": "object"
                    },
                    "conduccion": {
                        "type": "string",
                        "enum": [
                            "izquierda",
                            "derecha"
                        ],
                        "example": "izquierda",
                        "nullable": true
                    },
                    "idiomas": {
                        "description": "Idiomas hablados en el país",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "nombre": {
                                    "type": "string",
                                    "example": "Español"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "lugares": {
                        "description": "Lugares turísticos del país",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 5
                                },
                                "nombre": {
                                    "type": "string",
                                    "example": "Cataratas del Iguazú"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "PaisDB": {
                "title": "Pais (BBDD)",
                "description": "Esquema basado en la tabla de base de datos paises",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Argentina"
                    },
                    "continente_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "capital": {
                        "type": "string",
                        "example": "Buenos Aires"
                    },
                    "bandera_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/bandera.png",
                        "nullable": true
                    },
                    "conduccion": {
                        "type": "string",
                        "enum": [
                            "izquierda",
                            "derecha"
                        ],
                        "example": "izquierda",
                        "nullable": true
                    },
                    "idiomas": {
                        "description": "IDs de los idiomas del país",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}